Extract switchlink_db_types from switchlink_db
- Extracted switch_db_types.h from switchlink_db.h.
- Defined switchlink_db_types bazel target.
- Defined switchlink_db_int bazel target.
