+++ switchlink/BUILD.bazel
@@ -55,12 +55,11 @@ cc_test(
 
 krnlmon_cc_library(
     name = "switchlink_db",
-    srcs = [
-        "switchlink_db.c",
-        "switchlink_db.h",
-        "switchlink_db_int.h",
-    ],
+    srcs = ["switchlink_db.c"],
+    hdrs = ["switchlink_db.h"],
     deps = [
+        ":switchlink_db_int",
+        ":switchlink_db_types",
         ":switchlink_defs",
         ":switchlink_int",
         ":switchlink_link_types",
@@ -71,18 +70,38 @@ krnlmon_cc_library(
 )
 
 krnlmon_cc_library(
-    name = "switchlink_globals",
-    srcs = ["switchlink_globals.c"],
-    hdrs = ["switchlink_globals.h"],
-    deps = [":switchlink_defs"],
+    name = "switchlink_db_int",
+    hdrs = ["switchlink_db_int.h"],
+    deps = [
+        ":switchlink_db_types",
+        "@target_utils//:tommyds",
+    ],
+)
+
+krnlmon_cc_library(
+    name = "switchlink_db_types",
+    hdrs = ["switchlink_db_types.h"],
+    deps = [
+        ":switchlink_link_types",
+        ":switchlink_types",
+    ],
 )
 
 krnlmon_cc_library(
     name = "switchlink_defs",
-    srcs = ["switchlink_defs.h"],
+    hdrs = ["switchlink_defs.h"],
     deps = [
         ":switchlink_types",
         "//switchutils:switch_utils",
+    ],
+)
+
+krnlmon_cc_library(
+    name = "switchlink_globals",
+    srcs = ["switchlink_globals.c"],
+    hdrs = ["switchlink_globals.h"],
+    deps = [
+        ":switchlink_defs",
     ],
 )
 
@@ -159,6 +178,7 @@ krnlmon_cc_library(
     name = "switchlink_neigh",
     srcs = ["switchlink_neigh.c"],
     deps = [
+        ":switchlink_db",
         ":switchlink_defs",
         ":switchlink_globals",
         ":switchlink_handlers",
