+++ switchlink/switchlink_neigh.c
@@ -20,6 +20,7 @@
 #include <net/if.h>
 #include <netlink/msg.h>
 
+#include "switchlink_db.h"
 #include "switchlink_defs.h"
 #include "switchlink_globals.h"
 #include "switchlink_handlers.h"
